Moves CloseSign component close to the Overlay one, as they are strongly linked
